DankGasai ALPHA V2.2
- Added admin system. -addadmin %user%
- Added -joinserver which returns my invite url
- Added -kick %user% so you can kick a user from your server
- Added -mute %user% and abilities to mute users. Also stops user from
using bot commands
- Forgot to remove raw.java ._.
- Recoded restart.java to work with muting system along with every other
class
- Added reseting name, avatar and game with -reset
- Added -setgame to set my game
- Added -setname to set my name
- Added CommandsListener.java which logs commands and sends to Log
Channel
- Added MutedListener to delete messages if you are muted
- Added a few replies to replieslistener.java
- Removed all Meme classes and recoded entire Meme system
- Updated setCommands.java to work with new meme system
- Added -share %img% to share memes instead of -$img#
- Added -addimg %imgUrl% $extension% %name% for memes
- Added -share with no args reads back possible solutions/memes
- Added -roll %intger% to roll a number between 0 and %integer%
- Added -serverinfo which returns all info possible on current server
- Added -getinfo %user% to get some info about a user
- Deleted .setAdmins.java due to new admin system writing to txt file
- Recoded todo.java so now todo list is written and read from a todo.txt
file
- Cleaned up Main class to work with restarting on a VPS.
- Updated version in settings.java to "Alpha 2.2"
- Added dependencies folder into src folder for easy access
- Updated to Javacord 2.0.10-6

+++ src/Home/Administration/Shutdown.java
@@ -5,6 +5,8 @@ import Home.Settings;
 import de.btobastian.javacord.DiscordAPI;
 import de.btobastian.javacord.entities.message.Message;
 import de.btobastian.javacord.listener.message.MessageCreateListener;
+
+import static Home.Administration.Mute.muted;
 import static Home.Main.delay;
 import static Home.Main.*;
 public class Shutdown implements MessageCreateListener {
@@ -13,16 +15,18 @@ public class Shutdown implements MessageCreateListener {
         if(message.getContent().startsWith(Settings.getCommandStart() + "shutdown")){
             String[] args = message.getContent().split(" ");
             args[0] = args[0].replaceFirst(Settings.getCommandStart(), "");
-            if(args[0].equalsIgnoreCase("shutdown")){
-                message.delete();
-                    if(Main.admins.contains(message.getAuthor().getId())) {
+            if(args[0].equalsIgnoreCase("shutdown")) {
+                if (!muted.contains(message.getAuthor().getId())) {
+                    message.delete();
+                    if (Main.admins.contains(message.getAuthor().getId())) {
                         message.reply(Settings.getMsgStart() + "If....If you say so master.. " + message.getAuthor().getMentionTag() + ".. Shutting down.");
                         delay();
                         setShutdownNatural(true);
                         delay();
                         System.exit(0);
-                }else{
-                    message.reply(Settings.getAdminMsg());
+                    } else {
+                        message.reply(Settings.getAdminMsg());
+                    }
                 }
             }
         }
